requirements.txt
gitpython
requests
python-dotenv
urlextract